秒数取整

+++ Extension/AgeExtension.cs
@@ -4,49 +4,68 @@ namespace Extension
 {
     public static class AgeExtensions
     {
-
         private const int Second = 1;
         private const int Minute = 60 * Second;
         private const int Hour   = 60 * Minute;
         private const int Day    = 24 * Hour;
         private const int Month  = 30 * Day;
 
-        public static string ToFriendlyDisplay(this DateTime dateTime) {
+        public static string ToFriendlyDisplay(this DateTime dateTime)
+        {
             var ts    = DateTime.Now - dateTime;
             var delta = ts.TotalSeconds;
-            if (delta < 0) {
+            if (delta < 0)
+            {
                 return "not yet";
             }
-            if (delta < 1 * Minute) {
+
+            if (delta < 1 * Minute)
+            {
                 return ts.Seconds == 1 ? "1 second ago" : ts.Seconds + " seconds ago";
             }
-            if (delta < 2 * Minute) {
+
+            if (delta < 2 * Minute)
+            {
                 return "1 minute ago";
             }
-            if (delta < 45 * Minute) {
+
+            if (delta < 45 * Minute)
+            {
                 return ts.Minutes + "minute";
             }
-            if (delta < 90 * Minute) {
+
+            if (delta < 90 * Minute)
+            {
                 return "1 hour ago";
             }
-            if (delta < 24 * Hour) {
+
+            if (delta < 24 * Hour)
+            {
                 return ts.Hours + " hours ago";
             }
-            if (delta < 48 * Hour) {
+
+            if (delta < 48 * Hour)
+            {
                 return "yesterday";
             }
-            if (delta < 30 * Day) {
+
+            if (delta < 30 * Day)
+            {
                 return ts.Days + " days ago";
             }
-            if (delta < 12 * Month) {
+
+            if (delta < 12 * Month)
+            {
                 var months = Convert.ToInt32(Math.Floor((double)ts.Days / 30));
                 return months <= 1 ? "A month ago" : months + " months ago";
             }
-            else {
+            else
+            {
                 var years = Convert.ToInt32(Math.Floor((double)ts.Days / 365));
                 return years <= 1 ? "a year ago" : years + " years ago";
             }
         }
+
         /// <summary>
         /// 计算时间年龄
         /// </summary>
@@ -73,23 +92,25 @@ namespace Extension
             var dateTime = DateTime.Parse(time.ToString("yyyy-MM-dd'T'HH:mm:ssZ"));
             return dateTime.Age();
         }
+
         public static string AgeFromUtc(this DateTime? time)
         {
-            if (time==null)
+            if (time == null)
             {
                 return "";
             }
+
             var dateTime = DateTime.Parse(time?.ToString("yyyy-MM-dd'T'HH:mm:ssZ")!);
             return dateTime.Age();
         }
 
         public static string Age(this DateTime time, DateTime burnTime)
         {
-            var s = time.Subtract(burnTime).TotalSeconds;
+            var s = (int)time.Subtract(burnTime).TotalSeconds;
             var m = (int)time.Subtract(burnTime).TotalMinutes;
             var h = (int)time.Subtract(burnTime).TotalHours;
             var d = (int)time.Subtract(burnTime).TotalDays;
-            var y =(int) s / 60 / 60 / 24 / 365;
+            var y = (int)s / 60 / 60 / 24 / 365;
 
             var ys = time.Subtract(burnTime).Seconds;
             var ym = time.Subtract(burnTime).Minutes;
